requirements.txt
streamlit
pandas
spacy[lookups]
docx2txt
pdfminer.six
openpyxl
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz
